+++ lib/Board.rb
