+++ lib/Functions.rb
