+++ lib/SymbolTB.rb
